frontend: add a confirmation message for importing data

+++ tracker/frontend.js
@@ -612,6 +612,7 @@ importBtn.onclick = () => {
     if (eiinput.value.length <= 1) return eistatus.innerText = "import failed: data cannot be 0!"
     try { // safeguard just in case someone inputs fake data for the lolz
         JSON.parse(eiinput.value)
+        if (confirm("Importing this data is irreversible and any data you currently have will be removed. Are you sure you want to continue?") != true) return;
         try {
             localStorage.setItem(gamedatastore, eiinput.value)
             eistatus.innerHTML = `imported! <a href="${window.location.href}" style="color: white;">refresh</a> to see changes!`
